feat: 상품 등록/수정시 카테고리 매핑 가능하게 변경

+++ src/main/java/command/admin/category/CategoryCreateCommand.java
@@ -123,7 +123,7 @@ public class CategoryCreateCommand implements Command {
         // 상위 카테고리
         String parentIdStr = request.getParameter("parentId");
         if (parentIdStr != null && !parentIdStr.isEmpty()) {
-            dto.setParentId(Integer.parseInt(parentIdStr));
+            dto.setParentId(Long.parseLong(parentIdStr));
         }
 
         // 순서
